Cooccurrence analysis

Cooccurrence analysis.

Usage

ec_cooc(x, ...)

# S3 method for pa
ec_cooc(x, ...)

# S3 method for cooc_count
ec_cooc(x, nsit, ...)

ec_cooc_pairwise(x, ...)

# S3 method for pa
ec_cooc_pairwise(x, ...)

# S3 method for cooc_count
ec_cooc_pairwise(x, nsit, ...)

ec_cooc_species(x, ...)

# S3 method for pa
ec_cooc_species(x, ...)

# S3 method for cooc_count
ec_cooc_species(x, nsit, ...)

ec_cooc_global(x, ...)

# S3 method for pa
ec_cooc_global(x, ...)

# S3 method for cooc_count
ec_cooc_global(x, nsit, ...)

Arguments

x

object to be converted into an object class pa.

...

ignored.

nsit

number of sites.

Functions

ec_cooc_pairwise(): Return a matrix of cooccurrence metrics for pair of species.

ec_cooc_species(): Return a matrix of cooccurrence metrics for species.

ec_cooc_global(): Return a matrix of cooccurrence metrics for species.

Examples

mat0 <- matrix(0, 10, 10)
mat1 <- matrix(1, 10, 10)
matU <- rbind(cbind(mat1, mat0), cbind(mat0, mat1))
ec_cooc(ec_as_pa(matU))
#> $cooc_count
#>     species1 species2 case_10 case_01 case_11 case_00 case_spc1 case_spc2
#> 1     spc_01   spc_02       0       0      10      10        10        10
#> 2     spc_01   spc_03       0       0      10      10        10        10
#> 3     spc_01   spc_04       0       0      10      10        10        10
#> 4     spc_01   spc_05       0       0      10      10        10        10
#> 5     spc_01   spc_06       0       0      10      10        10        10
#> 6     spc_01   spc_07       0       0      10      10        10        10
#> 7     spc_01   spc_08       0       0      10      10        10        10
#> 8     spc_01   spc_09       0       0      10      10        10        10
#> 9     spc_01   spc_10       0       0      10      10        10        10
#> 10    spc_01   spc_11      10      10       0       0        10        10
#> 11    spc_01   spc_12      10      10       0       0        10        10
#> 12    spc_01   spc_13      10      10       0       0        10        10
#> 13    spc_01   spc_14      10      10       0       0        10        10
#> 14    spc_01   spc_15      10      10       0       0        10        10
#> 15    spc_01   spc_16      10      10       0       0        10        10
#> 16    spc_01   spc_17      10      10       0       0        10        10
#> 17    spc_01   spc_18      10      10       0       0        10        10
#> 18    spc_01   spc_19      10      10       0       0        10        10
#> 19    spc_01   spc_20      10      10       0       0        10        10
#> 20    spc_02   spc_03       0       0      10      10        10        10
#> 21    spc_02   spc_04       0       0      10      10        10        10
#> 22    spc_02   spc_05       0       0      10      10        10        10
#> 23    spc_02   spc_06       0       0      10      10        10        10
#> 24    spc_02   spc_07       0       0      10      10        10        10
#> 25    spc_02   spc_08       0       0      10      10        10        10
#> 26    spc_02   spc_09       0       0      10      10        10        10
#> 27    spc_02   spc_10       0       0      10      10        10        10
#> 28    spc_02   spc_11      10      10       0       0        10        10
#> 29    spc_02   spc_12      10      10       0       0        10        10
#> 30    spc_02   spc_13      10      10       0       0        10        10
#> 31    spc_02   spc_14      10      10       0       0        10        10
#> 32    spc_02   spc_15      10      10       0       0        10        10
#> 33    spc_02   spc_16      10      10       0       0        10        10
#> 34    spc_02   spc_17      10      10       0       0        10        10
#> 35    spc_02   spc_18      10      10       0       0        10        10
#> 36    spc_02   spc_19      10      10       0       0        10        10
#> 37    spc_02   spc_20      10      10       0       0        10        10
#> 38    spc_03   spc_04       0       0      10      10        10        10
#> 39    spc_03   spc_05       0       0      10      10        10        10
#> 40    spc_03   spc_06       0       0      10      10        10        10
#> 41    spc_03   spc_07       0       0      10      10        10        10
#> 42    spc_03   spc_08       0       0      10      10        10        10
#> 43    spc_03   spc_09       0       0      10      10        10        10
#> 44    spc_03   spc_10       0       0      10      10        10        10
#> 45    spc_03   spc_11      10      10       0       0        10        10
#> 46    spc_03   spc_12      10      10       0       0        10        10
#> 47    spc_03   spc_13      10      10       0       0        10        10
#> 48    spc_03   spc_14      10      10       0       0        10        10
#> 49    spc_03   spc_15      10      10       0       0        10        10
#> 50    spc_03   spc_16      10      10       0       0        10        10
#> 51    spc_03   spc_17      10      10       0       0        10        10
#> 52    spc_03   spc_18      10      10       0       0        10        10
#> 53    spc_03   spc_19      10      10       0       0        10        10
#> 54    spc_03   spc_20      10      10       0       0        10        10
#> 55    spc_04   spc_05       0       0      10      10        10        10
#> 56    spc_04   spc_06       0       0      10      10        10        10
#> 57    spc_04   spc_07       0       0      10      10        10        10
#> 58    spc_04   spc_08       0       0      10      10        10        10
#> 59    spc_04   spc_09       0       0      10      10        10        10
#> 60    spc_04   spc_10       0       0      10      10        10        10
#> 61    spc_04   spc_11      10      10       0       0        10        10
#> 62    spc_04   spc_12      10      10       0       0        10        10
#> 63    spc_04   spc_13      10      10       0       0        10        10
#> 64    spc_04   spc_14      10      10       0       0        10        10
#> 65    spc_04   spc_15      10      10       0       0        10        10
#> 66    spc_04   spc_16      10      10       0       0        10        10
#> 67    spc_04   spc_17      10      10       0       0        10        10
#> 68    spc_04   spc_18      10      10       0       0        10        10
#> 69    spc_04   spc_19      10      10       0       0        10        10
#> 70    spc_04   spc_20      10      10       0       0        10        10
#> 71    spc_05   spc_06       0       0      10      10        10        10
#> 72    spc_05   spc_07       0       0      10      10        10        10
#> 73    spc_05   spc_08       0       0      10      10        10        10
#> 74    spc_05   spc_09       0       0      10      10        10        10
#> 75    spc_05   spc_10       0       0      10      10        10        10
#> 76    spc_05   spc_11      10      10       0       0        10        10
#> 77    spc_05   spc_12      10      10       0       0        10        10
#> 78    spc_05   spc_13      10      10       0       0        10        10
#> 79    spc_05   spc_14      10      10       0       0        10        10
#> 80    spc_05   spc_15      10      10       0       0        10        10
#> 81    spc_05   spc_16      10      10       0       0        10        10
#> 82    spc_05   spc_17      10      10       0       0        10        10
#> 83    spc_05   spc_18      10      10       0       0        10        10
#> 84    spc_05   spc_19      10      10       0       0        10        10
#> 85    spc_05   spc_20      10      10       0       0        10        10
#> 86    spc_06   spc_07       0       0      10      10        10        10
#> 87    spc_06   spc_08       0       0      10      10        10        10
#> 88    spc_06   spc_09       0       0      10      10        10        10
#> 89    spc_06   spc_10       0       0      10      10        10        10
#> 90    spc_06   spc_11      10      10       0       0        10        10
#> 91    spc_06   spc_12      10      10       0       0        10        10
#> 92    spc_06   spc_13      10      10       0       0        10        10
#> 93    spc_06   spc_14      10      10       0       0        10        10
#> 94    spc_06   spc_15      10      10       0       0        10        10
#> 95    spc_06   spc_16      10      10       0       0        10        10
#> 96    spc_06   spc_17      10      10       0       0        10        10
#> 97    spc_06   spc_18      10      10       0       0        10        10
#> 98    spc_06   spc_19      10      10       0       0        10        10
#> 99    spc_06   spc_20      10      10       0       0        10        10
#> 100   spc_07   spc_08       0       0      10      10        10        10
#> 101   spc_07   spc_09       0       0      10      10        10        10
#> 102   spc_07   spc_10       0       0      10      10        10        10
#> 103   spc_07   spc_11      10      10       0       0        10        10
#> 104   spc_07   spc_12      10      10       0       0        10        10
#> 105   spc_07   spc_13      10      10       0       0        10        10
#> 106   spc_07   spc_14      10      10       0       0        10        10
#> 107   spc_07   spc_15      10      10       0       0        10        10
#> 108   spc_07   spc_16      10      10       0       0        10        10
#> 109   spc_07   spc_17      10      10       0       0        10        10
#> 110   spc_07   spc_18      10      10       0       0        10        10
#> 111   spc_07   spc_19      10      10       0       0        10        10
#> 112   spc_07   spc_20      10      10       0       0        10        10
#> 113   spc_08   spc_09       0       0      10      10        10        10
#> 114   spc_08   spc_10       0       0      10      10        10        10
#> 115   spc_08   spc_11      10      10       0       0        10        10
#> 116   spc_08   spc_12      10      10       0       0        10        10
#> 117   spc_08   spc_13      10      10       0       0        10        10
#> 118   spc_08   spc_14      10      10       0       0        10        10
#> 119   spc_08   spc_15      10      10       0       0        10        10
#> 120   spc_08   spc_16      10      10       0       0        10        10
#> 121   spc_08   spc_17      10      10       0       0        10        10
#> 122   spc_08   spc_18      10      10       0       0        10        10
#> 123   spc_08   spc_19      10      10       0       0        10        10
#> 124   spc_08   spc_20      10      10       0       0        10        10
#> 125   spc_09   spc_10       0       0      10      10        10        10
#> 126   spc_09   spc_11      10      10       0       0        10        10
#> 127   spc_09   spc_12      10      10       0       0        10        10
#> 128   spc_09   spc_13      10      10       0       0        10        10
#> 129   spc_09   spc_14      10      10       0       0        10        10
#> 130   spc_09   spc_15      10      10       0       0        10        10
#> 131   spc_09   spc_16      10      10       0       0        10        10
#> 132   spc_09   spc_17      10      10       0       0        10        10
#> 133   spc_09   spc_18      10      10       0       0        10        10
#> 134   spc_09   spc_19      10      10       0       0        10        10
#> 135   spc_09   spc_20      10      10       0       0        10        10
#> 136   spc_10   spc_11      10      10       0       0        10        10
#> 137   spc_10   spc_12      10      10       0       0        10        10
#> 138   spc_10   spc_13      10      10       0       0        10        10
#> 139   spc_10   spc_14      10      10       0       0        10        10
#> 140   spc_10   spc_15      10      10       0       0        10        10
#> 141   spc_10   spc_16      10      10       0       0        10        10
#> 142   spc_10   spc_17      10      10       0       0        10        10
#> 143   spc_10   spc_18      10      10       0       0        10        10
#> 144   spc_10   spc_19      10      10       0       0        10        10
#> 145   spc_10   spc_20      10      10       0       0        10        10
#> 146   spc_11   spc_12       0       0      10      10        10        10
#> 147   spc_11   spc_13       0       0      10      10        10        10
#> 148   spc_11   spc_14       0       0      10      10        10        10
#> 149   spc_11   spc_15       0       0      10      10        10        10
#> 150   spc_11   spc_16       0       0      10      10        10        10
#> 151   spc_11   spc_17       0       0      10      10        10        10
#> 152   spc_11   spc_18       0       0      10      10        10        10
#> 153   spc_11   spc_19       0       0      10      10        10        10
#> 154   spc_11   spc_20       0       0      10      10        10        10
#> 155   spc_12   spc_13       0       0      10      10        10        10
#> 156   spc_12   spc_14       0       0      10      10        10        10
#> 157   spc_12   spc_15       0       0      10      10        10        10
#> 158   spc_12   spc_16       0       0      10      10        10        10
#> 159   spc_12   spc_17       0       0      10      10        10        10
#> 160   spc_12   spc_18       0       0      10      10        10        10
#> 161   spc_12   spc_19       0       0      10      10        10        10
#> 162   spc_12   spc_20       0       0      10      10        10        10
#> 163   spc_13   spc_14       0       0      10      10        10        10
#> 164   spc_13   spc_15       0       0      10      10        10        10
#> 165   spc_13   spc_16       0       0      10      10        10        10
#> 166   spc_13   spc_17       0       0      10      10        10        10
#> 167   spc_13   spc_18       0       0      10      10        10        10
#> 168   spc_13   spc_19       0       0      10      10        10        10
#> 169   spc_13   spc_20       0       0      10      10        10        10
#> 170   spc_14   spc_15       0       0      10      10        10        10
#> 171   spc_14   spc_16       0       0      10      10        10        10
#> 172   spc_14   spc_17       0       0      10      10        10        10
#> 173   spc_14   spc_18       0       0      10      10        10        10
#> 174   spc_14   spc_19       0       0      10      10        10        10
#> 175   spc_14   spc_20       0       0      10      10        10        10
#> 176   spc_15   spc_16       0       0      10      10        10        10
#> 177   spc_15   spc_17       0       0      10      10        10        10
#> 178   spc_15   spc_18       0       0      10      10        10        10
#> 179   spc_15   spc_19       0       0      10      10        10        10
#> 180   spc_15   spc_20       0       0      10      10        10        10
#> 181   spc_16   spc_17       0       0      10      10        10        10
#> 182   spc_16   spc_18       0       0      10      10        10        10
#> 183   spc_16   spc_19       0       0      10      10        10        10
#> 184   spc_16   spc_20       0       0      10      10        10        10
#> 185   spc_17   spc_18       0       0      10      10        10        10
#> 186   spc_17   spc_19       0       0      10      10        10        10
#> 187   spc_17   spc_20       0       0      10      10        10        10
#> 188   spc_18   spc_19       0       0      10      10        10        10
#> 189   spc_18   spc_20       0       0      10      10        10        10
#> 190   spc_19   spc_20       0       0      10      10        10        10
#> 
#> $pairwise
#>     species1 species2     zs_bi     zs_hy binary_covariance mean_pairwise_index
#> 1     spc_01   spc_02  2.581989  4.358899         0.2631579                 Inf
#> 2     spc_01   spc_03  2.581989  4.358899         0.2631579                 Inf
#> 3     spc_01   spc_04  2.581989  4.358899         0.2631579                 Inf
#> 4     spc_01   spc_05  2.581989  4.358899         0.2631579                 Inf
#> 5     spc_01   spc_06  2.581989  4.358899         0.2631579                 Inf
#> 6     spc_01   spc_07  2.581989  4.358899         0.2631579                 Inf
#> 7     spc_01   spc_08  2.581989  4.358899         0.2631579                 Inf
#> 8     spc_01   spc_09  2.581989  4.358899         0.2631579                 Inf
#> 9     spc_01   spc_10  2.581989  4.358899         0.2631579                 Inf
#> 10    spc_01   spc_11 -2.581989 -4.358899        -0.2631579                   0
#> 11    spc_01   spc_12 -2.581989 -4.358899        -0.2631579                   0
#> 12    spc_01   spc_13 -2.581989 -4.358899        -0.2631579                   0
#> 13    spc_01   spc_14 -2.581989 -4.358899        -0.2631579                   0
#> 14    spc_01   spc_15 -2.581989 -4.358899        -0.2631579                   0
#> 15    spc_01   spc_16 -2.581989 -4.358899        -0.2631579                   0
#> 16    spc_01   spc_17 -2.581989 -4.358899        -0.2631579                   0
#> 17    spc_01   spc_18 -2.581989 -4.358899        -0.2631579                   0
#> 18    spc_01   spc_19 -2.581989 -4.358899        -0.2631579                   0
#> 19    spc_01   spc_20 -2.581989 -4.358899        -0.2631579                   0
#> 20    spc_02   spc_03  2.581989  4.358899         0.2631579                 Inf
#> 21    spc_02   spc_04  2.581989  4.358899         0.2631579                 Inf
#> 22    spc_02   spc_05  2.581989  4.358899         0.2631579                 Inf
#> 23    spc_02   spc_06  2.581989  4.358899         0.2631579                 Inf
#> 24    spc_02   spc_07  2.581989  4.358899         0.2631579                 Inf
#> 25    spc_02   spc_08  2.581989  4.358899         0.2631579                 Inf
#> 26    spc_02   spc_09  2.581989  4.358899         0.2631579                 Inf
#> 27    spc_02   spc_10  2.581989  4.358899         0.2631579                 Inf
#> 28    spc_02   spc_11 -2.581989 -4.358899        -0.2631579                   0
#> 29    spc_02   spc_12 -2.581989 -4.358899        -0.2631579                   0
#> 30    spc_02   spc_13 -2.581989 -4.358899        -0.2631579                   0
#> 31    spc_02   spc_14 -2.581989 -4.358899        -0.2631579                   0
#> 32    spc_02   spc_15 -2.581989 -4.358899        -0.2631579                   0
#> 33    spc_02   spc_16 -2.581989 -4.358899        -0.2631579                   0
#> 34    spc_02   spc_17 -2.581989 -4.358899        -0.2631579                   0
#> 35    spc_02   spc_18 -2.581989 -4.358899        -0.2631579                   0
#> 36    spc_02   spc_19 -2.581989 -4.358899        -0.2631579                   0
#> 37    spc_02   spc_20 -2.581989 -4.358899        -0.2631579                   0
#> 38    spc_03   spc_04  2.581989  4.358899         0.2631579                 Inf
#> 39    spc_03   spc_05  2.581989  4.358899         0.2631579                 Inf
#> 40    spc_03   spc_06  2.581989  4.358899         0.2631579                 Inf
#> 41    spc_03   spc_07  2.581989  4.358899         0.2631579                 Inf
#> 42    spc_03   spc_08  2.581989  4.358899         0.2631579                 Inf
#> 43    spc_03   spc_09  2.581989  4.358899         0.2631579                 Inf
#> 44    spc_03   spc_10  2.581989  4.358899         0.2631579                 Inf
#> 45    spc_03   spc_11 -2.581989 -4.358899        -0.2631579                   0
#> 46    spc_03   spc_12 -2.581989 -4.358899        -0.2631579                   0
#> 47    spc_03   spc_13 -2.581989 -4.358899        -0.2631579                   0
#> 48    spc_03   spc_14 -2.581989 -4.358899        -0.2631579                   0
#> 49    spc_03   spc_15 -2.581989 -4.358899        -0.2631579                   0
#> 50    spc_03   spc_16 -2.581989 -4.358899        -0.2631579                   0
#> 51    spc_03   spc_17 -2.581989 -4.358899        -0.2631579                   0
#> 52    spc_03   spc_18 -2.581989 -4.358899        -0.2631579                   0
#> 53    spc_03   spc_19 -2.581989 -4.358899        -0.2631579                   0
#> 54    spc_03   spc_20 -2.581989 -4.358899        -0.2631579                   0
#> 55    spc_04   spc_05  2.581989  4.358899         0.2631579                 Inf
#> 56    spc_04   spc_06  2.581989  4.358899         0.2631579                 Inf
#> 57    spc_04   spc_07  2.581989  4.358899         0.2631579                 Inf
#> 58    spc_04   spc_08  2.581989  4.358899         0.2631579                 Inf
#> 59    spc_04   spc_09  2.581989  4.358899         0.2631579                 Inf
#> 60    spc_04   spc_10  2.581989  4.358899         0.2631579                 Inf
#> 61    spc_04   spc_11 -2.581989 -4.358899        -0.2631579                   0
#> 62    spc_04   spc_12 -2.581989 -4.358899        -0.2631579                   0
#> 63    spc_04   spc_13 -2.581989 -4.358899        -0.2631579                   0
#> 64    spc_04   spc_14 -2.581989 -4.358899        -0.2631579                   0
#> 65    spc_04   spc_15 -2.581989 -4.358899        -0.2631579                   0
#> 66    spc_04   spc_16 -2.581989 -4.358899        -0.2631579                   0
#> 67    spc_04   spc_17 -2.581989 -4.358899        -0.2631579                   0
#> 68    spc_04   spc_18 -2.581989 -4.358899        -0.2631579                   0
#> 69    spc_04   spc_19 -2.581989 -4.358899        -0.2631579                   0
#> 70    spc_04   spc_20 -2.581989 -4.358899        -0.2631579                   0
#> 71    spc_05   spc_06  2.581989  4.358899         0.2631579                 Inf
#> 72    spc_05   spc_07  2.581989  4.358899         0.2631579                 Inf
#> 73    spc_05   spc_08  2.581989  4.358899         0.2631579                 Inf
#> 74    spc_05   spc_09  2.581989  4.358899         0.2631579                 Inf
#> 75    spc_05   spc_10  2.581989  4.358899         0.2631579                 Inf
#> 76    spc_05   spc_11 -2.581989 -4.358899        -0.2631579                   0
#> 77    spc_05   spc_12 -2.581989 -4.358899        -0.2631579                   0
#> 78    spc_05   spc_13 -2.581989 -4.358899        -0.2631579                   0
#> 79    spc_05   spc_14 -2.581989 -4.358899        -0.2631579                   0
#> 80    spc_05   spc_15 -2.581989 -4.358899        -0.2631579                   0
#> 81    spc_05   spc_16 -2.581989 -4.358899        -0.2631579                   0
#> 82    spc_05   spc_17 -2.581989 -4.358899        -0.2631579                   0
#> 83    spc_05   spc_18 -2.581989 -4.358899        -0.2631579                   0
#> 84    spc_05   spc_19 -2.581989 -4.358899        -0.2631579                   0
#> 85    spc_05   spc_20 -2.581989 -4.358899        -0.2631579                   0
#> 86    spc_06   spc_07  2.581989  4.358899         0.2631579                 Inf
#> 87    spc_06   spc_08  2.581989  4.358899         0.2631579                 Inf
#> 88    spc_06   spc_09  2.581989  4.358899         0.2631579                 Inf
#> 89    spc_06   spc_10  2.581989  4.358899         0.2631579                 Inf
#> 90    spc_06   spc_11 -2.581989 -4.358899        -0.2631579                   0
#> 91    spc_06   spc_12 -2.581989 -4.358899        -0.2631579                   0
#> 92    spc_06   spc_13 -2.581989 -4.358899        -0.2631579                   0
#> 93    spc_06   spc_14 -2.581989 -4.358899        -0.2631579                   0
#> 94    spc_06   spc_15 -2.581989 -4.358899        -0.2631579                   0
#> 95    spc_06   spc_16 -2.581989 -4.358899        -0.2631579                   0
#> 96    spc_06   spc_17 -2.581989 -4.358899        -0.2631579                   0
#> 97    spc_06   spc_18 -2.581989 -4.358899        -0.2631579                   0
#> 98    spc_06   spc_19 -2.581989 -4.358899        -0.2631579                   0
#> 99    spc_06   spc_20 -2.581989 -4.358899        -0.2631579                   0
#> 100   spc_07   spc_08  2.581989  4.358899         0.2631579                 Inf
#> 101   spc_07   spc_09  2.581989  4.358899         0.2631579                 Inf
#> 102   spc_07   spc_10  2.581989  4.358899         0.2631579                 Inf
#> 103   spc_07   spc_11 -2.581989 -4.358899        -0.2631579                   0
#> 104   spc_07   spc_12 -2.581989 -4.358899        -0.2631579                   0
#> 105   spc_07   spc_13 -2.581989 -4.358899        -0.2631579                   0
#> 106   spc_07   spc_14 -2.581989 -4.358899        -0.2631579                   0
#> 107   spc_07   spc_15 -2.581989 -4.358899        -0.2631579                   0
#> 108   spc_07   spc_16 -2.581989 -4.358899        -0.2631579                   0
#> 109   spc_07   spc_17 -2.581989 -4.358899        -0.2631579                   0
#> 110   spc_07   spc_18 -2.581989 -4.358899        -0.2631579                   0
#> 111   spc_07   spc_19 -2.581989 -4.358899        -0.2631579                   0
#> 112   spc_07   spc_20 -2.581989 -4.358899        -0.2631579                   0
#> 113   spc_08   spc_09  2.581989  4.358899         0.2631579                 Inf
#> 114   spc_08   spc_10  2.581989  4.358899         0.2631579                 Inf
#> 115   spc_08   spc_11 -2.581989 -4.358899        -0.2631579                   0
#> 116   spc_08   spc_12 -2.581989 -4.358899        -0.2631579                   0
#> 117   spc_08   spc_13 -2.581989 -4.358899        -0.2631579                   0
#> 118   spc_08   spc_14 -2.581989 -4.358899        -0.2631579                   0
#> 119   spc_08   spc_15 -2.581989 -4.358899        -0.2631579                   0
#> 120   spc_08   spc_16 -2.581989 -4.358899        -0.2631579                   0
#> 121   spc_08   spc_17 -2.581989 -4.358899        -0.2631579                   0
#> 122   spc_08   spc_18 -2.581989 -4.358899        -0.2631579                   0
#> 123   spc_08   spc_19 -2.581989 -4.358899        -0.2631579                   0
#> 124   spc_08   spc_20 -2.581989 -4.358899        -0.2631579                   0
#> 125   spc_09   spc_10  2.581989  4.358899         0.2631579                 Inf
#> 126   spc_09   spc_11 -2.581989 -4.358899        -0.2631579                   0
#> 127   spc_09   spc_12 -2.581989 -4.358899        -0.2631579                   0
#> 128   spc_09   spc_13 -2.581989 -4.358899        -0.2631579                   0
#> 129   spc_09   spc_14 -2.581989 -4.358899        -0.2631579                   0
#> 130   spc_09   spc_15 -2.581989 -4.358899        -0.2631579                   0
#> 131   spc_09   spc_16 -2.581989 -4.358899        -0.2631579                   0
#> 132   spc_09   spc_17 -2.581989 -4.358899        -0.2631579                   0
#> 133   spc_09   spc_18 -2.581989 -4.358899        -0.2631579                   0
#> 134   spc_09   spc_19 -2.581989 -4.358899        -0.2631579                   0
#> 135   spc_09   spc_20 -2.581989 -4.358899        -0.2631579                   0
#> 136   spc_10   spc_11 -2.581989 -4.358899        -0.2631579                   0
#> 137   spc_10   spc_12 -2.581989 -4.358899        -0.2631579                   0
#> 138   spc_10   spc_13 -2.581989 -4.358899        -0.2631579                   0
#> 139   spc_10   spc_14 -2.581989 -4.358899        -0.2631579                   0
#> 140   spc_10   spc_15 -2.581989 -4.358899        -0.2631579                   0
#> 141   spc_10   spc_16 -2.581989 -4.358899        -0.2631579                   0
#> 142   spc_10   spc_17 -2.581989 -4.358899        -0.2631579                   0
#> 143   spc_10   spc_18 -2.581989 -4.358899        -0.2631579                   0
#> 144   spc_10   spc_19 -2.581989 -4.358899        -0.2631579                   0
#> 145   spc_10   spc_20 -2.581989 -4.358899        -0.2631579                   0
#> 146   spc_11   spc_12  2.581989  4.358899         0.2631579                 Inf
#> 147   spc_11   spc_13  2.581989  4.358899         0.2631579                 Inf
#> 148   spc_11   spc_14  2.581989  4.358899         0.2631579                 Inf
#> 149   spc_11   spc_15  2.581989  4.358899         0.2631579                 Inf
#> 150   spc_11   spc_16  2.581989  4.358899         0.2631579                 Inf
#> 151   spc_11   spc_17  2.581989  4.358899         0.2631579                 Inf
#> 152   spc_11   spc_18  2.581989  4.358899         0.2631579                 Inf
#> 153   spc_11   spc_19  2.581989  4.358899         0.2631579                 Inf
#> 154   spc_11   spc_20  2.581989  4.358899         0.2631579                 Inf
#> 155   spc_12   spc_13  2.581989  4.358899         0.2631579                 Inf
#> 156   spc_12   spc_14  2.581989  4.358899         0.2631579                 Inf
#> 157   spc_12   spc_15  2.581989  4.358899         0.2631579                 Inf
#> 158   spc_12   spc_16  2.581989  4.358899         0.2631579                 Inf
#> 159   spc_12   spc_17  2.581989  4.358899         0.2631579                 Inf
#> 160   spc_12   spc_18  2.581989  4.358899         0.2631579                 Inf
#> 161   spc_12   spc_19  2.581989  4.358899         0.2631579                 Inf
#> 162   spc_12   spc_20  2.581989  4.358899         0.2631579                 Inf
#> 163   spc_13   spc_14  2.581989  4.358899         0.2631579                 Inf
#> 164   spc_13   spc_15  2.581989  4.358899         0.2631579                 Inf
#> 165   spc_13   spc_16  2.581989  4.358899         0.2631579                 Inf
#> 166   spc_13   spc_17  2.581989  4.358899         0.2631579                 Inf
#> 167   spc_13   spc_18  2.581989  4.358899         0.2631579                 Inf
#> 168   spc_13   spc_19  2.581989  4.358899         0.2631579                 Inf
#> 169   spc_13   spc_20  2.581989  4.358899         0.2631579                 Inf
#> 170   spc_14   spc_15  2.581989  4.358899         0.2631579                 Inf
#> 171   spc_14   spc_16  2.581989  4.358899         0.2631579                 Inf
#> 172   spc_14   spc_17  2.581989  4.358899         0.2631579                 Inf
#> 173   spc_14   spc_18  2.581989  4.358899         0.2631579                 Inf
#> 174   spc_14   spc_19  2.581989  4.358899         0.2631579                 Inf
#> 175   spc_14   spc_20  2.581989  4.358899         0.2631579                 Inf
#> 176   spc_15   spc_16  2.581989  4.358899         0.2631579                 Inf
#> 177   spc_15   spc_17  2.581989  4.358899         0.2631579                 Inf
#> 178   spc_15   spc_18  2.581989  4.358899         0.2631579                 Inf
#> 179   spc_15   spc_19  2.581989  4.358899         0.2631579                 Inf
#> 180   spc_15   spc_20  2.581989  4.358899         0.2631579                 Inf
#> 181   spc_16   spc_17  2.581989  4.358899         0.2631579                 Inf
#> 182   spc_16   spc_18  2.581989  4.358899         0.2631579                 Inf
#> 183   spc_16   spc_19  2.581989  4.358899         0.2631579                 Inf
#> 184   spc_16   spc_20  2.581989  4.358899         0.2631579                 Inf
#> 185   spc_17   spc_18  2.581989  4.358899         0.2631579                 Inf
#> 186   spc_17   spc_19  2.581989  4.358899         0.2631579                 Inf
#> 187   spc_17   spc_20  2.581989  4.358899         0.2631579                 Inf
#> 188   spc_18   spc_19  2.581989  4.358899         0.2631579                 Inf
#> 189   spc_18   spc_20  2.581989  4.358899         0.2631579                 Inf
#> 190   spc_19   spc_20  2.581989  4.358899         0.2631579                 Inf
#>     mutual_information c_score_unit overlap_1_2 overlap_2_1 symmetry
#> 1               0.0025            0           1           1        0
#> 2               0.0025            0           1           1        0
#> 3               0.0025            0           1           1        0
#> 4               0.0025            0           1           1        0
#> 5               0.0025            0           1           1        0
#> 6               0.0025            0           1           1        0
#> 7               0.0025            0           1           1        0
#> 8               0.0025            0           1           1        0
#> 9               0.0025            0           1           1        0
#> 10              0.0025          100           0           0      NaN
#> 11              0.0025          100           0           0      NaN
#> 12              0.0025          100           0           0      NaN
#> 13              0.0025          100           0           0      NaN
#> 14              0.0025          100           0           0      NaN
#> 15              0.0025          100           0           0      NaN
#> 16              0.0025          100           0           0      NaN
#> 17              0.0025          100           0           0      NaN
#> 18              0.0025          100           0           0      NaN
#> 19              0.0025          100           0           0      NaN
#> 20              0.0025            0           1           1        0
#> 21              0.0025            0           1           1        0
#> 22              0.0025            0           1           1        0
#> 23              0.0025            0           1           1        0
#> 24              0.0025            0           1           1        0
#> 25              0.0025            0           1           1        0
#> 26              0.0025            0           1           1        0
#> 27              0.0025            0           1           1        0
#> 28              0.0025          100           0           0      NaN
#> 29              0.0025          100           0           0      NaN
#> 30              0.0025          100           0           0      NaN
#> 31              0.0025          100           0           0      NaN
#> 32              0.0025          100           0           0      NaN
#> 33              0.0025          100           0           0      NaN
#> 34              0.0025          100           0           0      NaN
#> 35              0.0025          100           0           0      NaN
#> 36              0.0025          100           0           0      NaN
#> 37              0.0025          100           0           0      NaN
#> 38              0.0025            0           1           1        0
#> 39              0.0025            0           1           1        0
#> 40              0.0025            0           1           1        0
#> 41              0.0025            0           1           1        0
#> 42              0.0025            0           1           1        0
#> 43              0.0025            0           1           1        0
#> 44              0.0025            0           1           1        0
#> 45              0.0025          100           0           0      NaN
#> 46              0.0025          100           0           0      NaN
#> 47              0.0025          100           0           0      NaN
#> 48              0.0025          100           0           0      NaN
#> 49              0.0025          100           0           0      NaN
#> 50              0.0025          100           0           0      NaN
#> 51              0.0025          100           0           0      NaN
#> 52              0.0025          100           0           0      NaN
#> 53              0.0025          100           0           0      NaN
#> 54              0.0025          100           0           0      NaN
#> 55              0.0025            0           1           1        0
#> 56              0.0025            0           1           1        0
#> 57              0.0025            0           1           1        0
#> 58              0.0025            0           1           1        0
#> 59              0.0025            0           1           1        0
#> 60              0.0025            0           1           1        0
#> 61              0.0025          100           0           0      NaN
#> 62              0.0025          100           0           0      NaN
#> 63              0.0025          100           0           0      NaN
#> 64              0.0025          100           0           0      NaN
#> 65              0.0025          100           0           0      NaN
#> 66              0.0025          100           0           0      NaN
#> 67              0.0025          100           0           0      NaN
#> 68              0.0025          100           0           0      NaN
#> 69              0.0025          100           0           0      NaN
#> 70              0.0025          100           0           0      NaN
#> 71              0.0025            0           1           1        0
#> 72              0.0025            0           1           1        0
#> 73              0.0025            0           1           1        0
#> 74              0.0025            0           1           1        0
#> 75              0.0025            0           1           1        0
#> 76              0.0025          100           0           0      NaN
#> 77              0.0025          100           0           0      NaN
#> 78              0.0025          100           0           0      NaN
#> 79              0.0025          100           0           0      NaN
#> 80              0.0025          100           0           0      NaN
#> 81              0.0025          100           0           0      NaN
#> 82              0.0025          100           0           0      NaN
#> 83              0.0025          100           0           0      NaN
#> 84              0.0025          100           0           0      NaN
#> 85              0.0025          100           0           0      NaN
#> 86              0.0025            0           1           1        0
#> 87              0.0025            0           1           1        0
#> 88              0.0025            0           1           1        0
#> 89              0.0025            0           1           1        0
#> 90              0.0025          100           0           0      NaN
#> 91              0.0025          100           0           0      NaN
#> 92              0.0025          100           0           0      NaN
#> 93              0.0025          100           0           0      NaN
#> 94              0.0025          100           0           0      NaN
#> 95              0.0025          100           0           0      NaN
#> 96              0.0025          100           0           0      NaN
#> 97              0.0025          100           0           0      NaN
#> 98              0.0025          100           0           0      NaN
#> 99              0.0025          100           0           0      NaN
#> 100             0.0025            0           1           1        0
#> 101             0.0025            0           1           1        0
#> 102             0.0025            0           1           1        0
#> 103             0.0025          100           0           0      NaN
#> 104             0.0025          100           0           0      NaN
#> 105             0.0025          100           0           0      NaN
#> 106             0.0025          100           0           0      NaN
#> 107             0.0025          100           0           0      NaN
#> 108             0.0025          100           0           0      NaN
#> 109             0.0025          100           0           0      NaN
#> 110             0.0025          100           0           0      NaN
#> 111             0.0025          100           0           0      NaN
#> 112             0.0025          100           0           0      NaN
#> 113             0.0025            0           1           1        0
#> 114             0.0025            0           1           1        0
#> 115             0.0025          100           0           0      NaN
#> 116             0.0025          100           0           0      NaN
#> 117             0.0025          100           0           0      NaN
#> 118             0.0025          100           0           0      NaN
#> 119             0.0025          100           0           0      NaN
#> 120             0.0025          100           0           0      NaN
#> 121             0.0025          100           0           0      NaN
#> 122             0.0025          100           0           0      NaN
#> 123             0.0025          100           0           0      NaN
#> 124             0.0025          100           0           0      NaN
#> 125             0.0025            0           1           1        0
#> 126             0.0025          100           0           0      NaN
#> 127             0.0025          100           0           0      NaN
#> 128             0.0025          100           0           0      NaN
#> 129             0.0025          100           0           0      NaN
#> 130             0.0025          100           0           0      NaN
#> 131             0.0025          100           0           0      NaN
#> 132             0.0025          100           0           0      NaN
#> 133             0.0025          100           0           0      NaN
#> 134             0.0025          100           0           0      NaN
#> 135             0.0025          100           0           0      NaN
#> 136             0.0025          100           0           0      NaN
#> 137             0.0025          100           0           0      NaN
#> 138             0.0025          100           0           0      NaN
#> 139             0.0025          100           0           0      NaN
#> 140             0.0025          100           0           0      NaN
#> 141             0.0025          100           0           0      NaN
#> 142             0.0025          100           0           0      NaN
#> 143             0.0025          100           0           0      NaN
#> 144             0.0025          100           0           0      NaN
#> 145             0.0025          100           0           0      NaN
#> 146             0.0025            0           1           1        0
#> 147             0.0025            0           1           1        0
#> 148             0.0025            0           1           1        0
#> 149             0.0025            0           1           1        0
#> 150             0.0025            0           1           1        0
#> 151             0.0025            0           1           1        0
#> 152             0.0025            0           1           1        0
#> 153             0.0025            0           1           1        0
#> 154             0.0025            0           1           1        0
#> 155             0.0025            0           1           1        0
#> 156             0.0025            0           1           1        0
#> 157             0.0025            0           1           1        0
#> 158             0.0025            0           1           1        0
#> 159             0.0025            0           1           1        0
#> 160             0.0025            0           1           1        0
#> 161             0.0025            0           1           1        0
#> 162             0.0025            0           1           1        0
#> 163             0.0025            0           1           1        0
#> 164             0.0025            0           1           1        0
#> 165             0.0025            0           1           1        0
#> 166             0.0025            0           1           1        0
#> 167             0.0025            0           1           1        0
#> 168             0.0025            0           1           1        0
#> 169             0.0025            0           1           1        0
#> 170             0.0025            0           1           1        0
#> 171             0.0025            0           1           1        0
#> 172             0.0025            0           1           1        0
#> 173             0.0025            0           1           1        0
#> 174             0.0025            0           1           1        0
#> 175             0.0025            0           1           1        0
#> 176             0.0025            0           1           1        0
#> 177             0.0025            0           1           1        0
#> 178             0.0025            0           1           1        0
#> 179             0.0025            0           1           1        0
#> 180             0.0025            0           1           1        0
#> 181             0.0025            0           1           1        0
#> 182             0.0025            0           1           1        0
#> 183             0.0025            0           1           1        0
#> 184             0.0025            0           1           1        0
#> 185             0.0025            0           1           1        0
#> 186             0.0025            0           1           1        0
#> 187             0.0025            0           1           1        0
#> 188             0.0025            0           1           1        0
#> 189             0.0025            0           1           1        0
#> 190             0.0025            0           1           1        0
#> 
#> $species
#>    species presence entropy robustness sensitivity
#> 1   spc_01      0.5       1          9           9
#> 2   spc_02      0.5       1          9           9
#> 3   spc_03      0.5       1          9           9
#> 4   spc_04      0.5       1          9           9
#> 5   spc_05      0.5       1          9           9
#> 6   spc_06      0.5       1          9           9
#> 7   spc_07      0.5       1          9           9
#> 8   spc_08      0.5       1          9           9
#> 9   spc_09      0.5       1          9           9
#> 10  spc_10      0.5       1          9           9
#> 11  spc_11      0.5       1          9           9
#> 12  spc_12      0.5       1          9           9
#> 13  spc_13      0.5       1          9           9
#> 14  spc_14      0.5       1          9           9
#> 15  spc_15      0.5       1          9           9
#> 16  spc_16      0.5       1          9           9
#> 17  spc_17      0.5       1          9           9
#> 18  spc_18      0.5       1          9           9
#> 19  spc_19      0.5       1          9           9
#> 20  spc_20      0.5       1          9           9
#> 
#> $global
#>    c_score c_score_S2
#> 1 52.63158   5263.158
#>