Compute beta diversity

Compute the beta diversity between all pairs of sites for a specific presence-absence matrix.

ec_betadiversity(x, methods = "bc")

Arguments

x
a pa object or an R object to a coerced to one (see ec_as_pa()).

methods
a vector of two-letters strings describing the methods te be used. Values should be taken among ra, bc, wi and ja (see details).

Value

A matrix with all the combinaisons of sites and the corresponding betadiversity.

Details

Currently ra stands for raw and returns the number of occurrence. Additional values are

bc: Bray-Curtis index,

wi: Wishart index,

ja: Jaccard index.

References

Legendre, P., and De Caceres M.. Beta Diversity as the Variance of Community Data: Dissimilarity Coefficients and Partitioning. Ecology Letters (2013).

Koleff, P., Gaston, K. J. & Lennon, J. J. Measuring beta diversity for presence-absence data. Journal of Animal Ecology (2003).

Examples

mat <- ec_generate_pa(20, 10, .4)
ec_betadiversity(mat)
#>      site1  site2         bc
#> 1   sit_01 sit_02 0.66666667
#> 2   sit_01 sit_03 0.50000000
#> 3   sit_01 sit_04 0.50000000
#> 4   sit_01 sit_05 0.45454545
#> 5   sit_01 sit_06 0.45454545
#> 6   sit_01 sit_07 0.33333333
#> 7   sit_01 sit_08 0.33333333
#> 8   sit_01 sit_09 0.66666667
#> 9   sit_01 sit_10 0.63636364
#> 10  sit_01 sit_11 0.50000000
#> 11  sit_01 sit_12 0.50000000
#> 12  sit_01 sit_13 0.75000000
#> 13  sit_01 sit_14 0.33333333
#> 14  sit_01 sit_15 0.55555556
#> 15  sit_01 sit_16 0.50000000
#> 16  sit_01 sit_17 0.40000000
#> 17  sit_01 sit_18 0.45454545
#> 18  sit_01 sit_19 0.71428571
#> 19  sit_01 sit_20 0.33333333
#> 20  sit_02 sit_03 0.33333333
#> 21  sit_02 sit_04 1.00000000
#> 22  sit_02 sit_05 0.45454545
#> 23  sit_02 sit_06 0.27272727
#> 24  sit_02 sit_07 0.77777778
#> 25  sit_02 sit_08 0.77777778
#> 26  sit_02 sit_09 0.33333333
#> 27  sit_02 sit_10 0.09090909
#> 28  sit_02 sit_11 0.16666667
#> 29  sit_02 sit_12 0.75000000
#> 30  sit_02 sit_13 0.75000000
#> 31  sit_02 sit_14 0.33333333
#> 32  sit_02 sit_15 0.55555556
#> 33  sit_02 sit_16 0.75000000
#> 34  sit_02 sit_17 0.60000000
#> 35  sit_02 sit_18 0.45454545
#> 36  sit_02 sit_19 1.00000000
#> 37  sit_02 sit_20 0.33333333
#> 38  sit_03 sit_04 0.75000000
#> 39  sit_03 sit_05 0.27272727
#> 40  sit_03 sit_06 0.45454545
#> 41  sit_03 sit_07 0.55555556
#> 42  sit_03 sit_08 0.77777778
#> 43  sit_03 sit_09 0.16666667
#> 44  sit_03 sit_10 0.27272727
#> 45  sit_03 sit_11 0.33333333
#> 46  sit_03 sit_12 1.00000000
#> 47  sit_03 sit_13 0.75000000
#> 48  sit_03 sit_14 0.50000000
#> 49  sit_03 sit_15 0.77777778
#> 50  sit_03 sit_16 0.50000000
#> 51  sit_03 sit_17 0.40000000
#> 52  sit_03 sit_18 0.45454545
#> 53  sit_03 sit_19 1.00000000
#> 54  sit_03 sit_20 0.33333333
#> 55  sit_04 sit_05 0.71428571
#> 56  sit_04 sit_06 0.71428571
#> 57  sit_04 sit_07 0.60000000
#> 58  sit_04 sit_08 0.20000000
#> 59  sit_04 sit_09 0.75000000
#> 60  sit_04 sit_10 1.00000000
#> 61  sit_04 sit_11 1.00000000
#> 62  sit_04 sit_12 0.50000000
#> 63  sit_04 sit_13 0.50000000
#> 64  sit_04 sit_14 0.75000000
#> 65  sit_04 sit_15 1.00000000
#> 66  sit_04 sit_16 0.50000000
#> 67  sit_04 sit_17 0.66666667
#> 68  sit_04 sit_18 0.71428571
#> 69  sit_04 sit_19 1.00000000
#> 70  sit_04 sit_20 0.75000000
#> 71  sit_05 sit_06 0.60000000
#> 72  sit_05 sit_07 0.75000000
#> 73  sit_05 sit_08 0.75000000
#> 74  sit_05 sit_09 0.45454545
#> 75  sit_05 sit_10 0.40000000
#> 76  sit_05 sit_11 0.45454545
#> 77  sit_05 sit_12 0.71428571
#> 78  sit_05 sit_13 0.42857143
#> 79  sit_05 sit_14 0.27272727
#> 80  sit_05 sit_15 0.75000000
#> 81  sit_05 sit_16 0.71428571
#> 82  sit_05 sit_17 0.11111111
#> 83  sit_05 sit_18 0.60000000
#> 84  sit_05 sit_19 1.00000000
#> 85  sit_05 sit_20 0.27272727
#> 86  sit_06 sit_07 0.50000000
#> 87  sit_06 sit_08 0.50000000
#> 88  sit_06 sit_09 0.45454545
#> 89  sit_06 sit_10 0.40000000
#> 90  sit_06 sit_11 0.45454545
#> 91  sit_06 sit_12 0.71428571
#> 92  sit_06 sit_13 1.00000000
#> 93  sit_06 sit_14 0.27272727
#> 94  sit_06 sit_15 0.50000000
#> 95  sit_06 sit_16 0.42857143
#> 96  sit_06 sit_17 0.55555556
#> 97  sit_06 sit_18 0.20000000
#> 98  sit_06 sit_19 1.00000000
#> 99  sit_06 sit_20 0.63636364
#> 100 sit_07 sit_08 0.66666667
#> 101 sit_07 sit_09 0.77777778
#> 102 sit_07 sit_10 0.75000000
#> 103 sit_07 sit_11 0.77777778
#> 104 sit_07 sit_12 1.00000000
#> 105 sit_07 sit_13 1.00000000
#> 106 sit_07 sit_14 0.77777778
#> 107 sit_07 sit_15 1.00000000
#> 108 sit_07 sit_16 0.20000000
#> 109 sit_07 sit_17 0.71428571
#> 110 sit_07 sit_18 0.75000000
#> 111 sit_07 sit_19 0.50000000
#> 112 sit_07 sit_20 0.77777778
#> 113 sit_08 sit_09 0.77777778
#> 114 sit_08 sit_10 0.75000000
#> 115 sit_08 sit_11 0.77777778
#> 116 sit_08 sit_12 0.20000000
#> 117 sit_08 sit_13 0.60000000
#> 118 sit_08 sit_14 0.55555556
#> 119 sit_08 sit_15 0.66666667
#> 120 sit_08 sit_16 0.60000000
#> 121 sit_08 sit_17 0.71428571
#> 122 sit_08 sit_18 0.50000000
#> 123 sit_08 sit_19 1.00000000
#> 124 sit_08 sit_20 0.55555556
#> 125 sit_09 sit_10 0.45454545
#> 126 sit_09 sit_11 0.33333333
#> 127 sit_09 sit_12 1.00000000
#> 128 sit_09 sit_13 0.75000000
#> 129 sit_09 sit_14 0.50000000
#> 130 sit_09 sit_15 0.55555556
#> 131 sit_09 sit_16 0.75000000
#> 132 sit_09 sit_17 0.60000000
#> 133 sit_09 sit_18 0.27272727
#> 134 sit_09 sit_19 1.00000000
#> 135 sit_09 sit_20 0.50000000
#> 136 sit_10 sit_11 0.27272727
#> 137 sit_10 sit_12 0.71428571
#> 138 sit_10 sit_13 0.71428571
#> 139 sit_10 sit_14 0.45454545
#> 140 sit_10 sit_15 0.75000000
#> 141 sit_10 sit_16 0.71428571
#> 142 sit_10 sit_17 0.55555556
#> 143 sit_10 sit_18 0.60000000
#> 144 sit_10 sit_19 1.00000000
#> 145 sit_10 sit_20 0.27272727
#> 146 sit_11 sit_12 0.75000000
#> 147 sit_11 sit_13 0.75000000
#> 148 sit_11 sit_14 0.33333333
#> 149 sit_11 sit_15 0.33333333
#> 150 sit_11 sit_16 0.75000000
#> 151 sit_11 sit_17 0.60000000
#> 152 sit_11 sit_18 0.45454545
#> 153 sit_11 sit_19 1.00000000
#> 154 sit_11 sit_20 0.16666667
#> 155 sit_12 sit_13 0.50000000
#> 156 sit_12 sit_14 0.50000000
#> 157 sit_12 sit_15 0.60000000
#> 158 sit_12 sit_16 1.00000000
#> 159 sit_12 sit_17 0.66666667
#> 160 sit_12 sit_18 0.71428571
#> 161 sit_12 sit_19 1.00000000
#> 162 sit_12 sit_20 0.50000000
#> 163 sit_13 sit_14 0.75000000
#> 164 sit_13 sit_15 1.00000000
#> 165 sit_13 sit_16 1.00000000
#> 166 sit_13 sit_17 0.66666667
#> 167 sit_13 sit_18 1.00000000
#> 168 sit_13 sit_19 1.00000000
#> 169 sit_13 sit_20 0.50000000
#> 170 sit_14 sit_15 0.33333333
#> 171 sit_14 sit_16 0.75000000
#> 172 sit_14 sit_17 0.20000000
#> 173 sit_14 sit_18 0.27272727
#> 174 sit_14 sit_19 1.00000000
#> 175 sit_14 sit_20 0.33333333
#> 176 sit_15 sit_16 1.00000000
#> 177 sit_15 sit_17 0.71428571
#> 178 sit_15 sit_18 0.25000000
#> 179 sit_15 sit_19 1.00000000
#> 180 sit_15 sit_20 0.55555556
#> 181 sit_16 sit_17 0.66666667
#> 182 sit_16 sit_18 0.71428571
#> 183 sit_16 sit_19 1.00000000
#> 184 sit_16 sit_20 0.75000000
#> 185 sit_17 sit_18 0.55555556
#> 186 sit_17 sit_19 1.00000000
#> 187 sit_17 sit_20 0.40000000
#> 188 sit_18 sit_19 1.00000000
#> 189 sit_18 sit_20 0.63636364
#> 190 sit_19 sit_20 1.00000000